Dialogue update
Now we can make it so that when the player interacts with an NPC or object the next time they interact with them new text will appear. This will hopefully liven up the overworld. As far as I know there isn't much of a limit to how many conversations one NPC can have, they only have to go one after the opther, have fun!